Break up the main plugin file
